userspace mailbox

+++ kernel/basic/arch/arm/lib/bcm283x/mailbox.c
@@ -1,11 +1,7 @@
-#include <kernel/system.h>
-#include <mm/kmalloc.h>
-#include <mm/mmu.h>
-#include <kstring.h>
 #include "mailbox.h"
 
 //void __attribute__((optimize("O0"))) mailbox_read(int channel, mail_message_t *msg) {
-void mailbox_read(int channel, mail_message_t *msg) {
+void mailbox_read(uint8_t channel, mail_message_t *msg) {
 	mail_status_t stat;
 
 	// Make sure that the message is from the right channel
@@ -21,7 +17,7 @@ void mailbox_read(int channel, mail_message_t *msg) {
 }
 
 //void __attribute__((optimize("O0"))) mailbox_send(int channel, mail_message_t* msg) {
-void  mailbox_send(int channel, mail_message_t* msg) {
+void  mailbox_send(uint8_t channel, mail_message_t* msg) {
 	mail_status_t stat;
 	msg->channel = channel;
